Crashing
Was working on the java doc. The objective was to connect the view to some test api and display it in list view. It crashes, I will try to redo it and if it working by next sunday. Feel free to look at it and tell me if you find any errors. I think I am going to end up redoing it from scratch.